+++ Coral.Native/Source/Coral/MethodInfo.cpp
@@ -29,7 +29,8 @@ namespace Coral {
 		{
 			int32_t parameterCount;
 			s_ManagedFunctions.GetMethodInfoParameterTypesFptr(m_Handle, nullptr, &parameterCount);
-			std::vector<TypeId> parameterTypes(parameterCount);
+
+			std::vector<TypeId> parameterTypes(static_cast<size_t>(parameterCount));
 			s_ManagedFunctions.GetMethodInfoParameterTypesFptr(m_Handle, parameterTypes.data(), &parameterCount);
 
 			m_ParameterTypes.resize(parameterTypes.size());
@@ -54,7 +55,8 @@ namespace Coral {
 	{
 		int32_t attributeCount;
 		s_ManagedFunctions.GetMethodInfoAttributesFptr(m_Handle, nullptr, &attributeCount);
-		std::vector<ManagedHandle> attributeHandles(attributeCount);
+
+		std::vector<ManagedHandle> attributeHandles(static_cast<size_t>(attributeCount));
 		s_ManagedFunctions.GetMethodInfoAttributesFptr(m_Handle, attributeHandles.data(), &attributeCount);
 
 		std::vector<Attribute> result(attributeHandles.size());
